+++ tpe_biblioteca_web2/Controller/LibroController.php
@@ -3,9 +3,9 @@
 require_once "./Model/LibroModel.php";
 require_once "./Model/AutorModel.php";
 require_once "./View/LibroView.php";
-require_once "./View/ErrorView.php";
 require_once "./View/AdminView.php";
 require_once "./Helpers/AuthHelper.php";
+require_once "./Helpers/FunctionHelper.php";
 
 class LibroController
 {
@@ -13,9 +13,9 @@ class LibroController
     private $model;
     private $view;
     private $viewAdmin;
-    private $viewError;
     private $authHelper;
     private $modelAutor;
+    private $functionHelper;
 
     function __construct()
     {
@@ -23,18 +23,30 @@ class LibroController
         $this->modelAutor = new AutorModel();
         $this->view = new LibroView();
         $this->viewAdmin = new AdminView();
-        $this->viewError = new ErrorView();
         $this->authHelper = new AuthHelper();
+        $this->functionHelper = new FunctionHelper();
     }
 
     function showHome(){
 
         $biblioteca = $this->model->getBiblioteca();
         if(!empty($biblioteca)){
-            $this->view->showBiblioteca($biblioteca);
+            if($this->authHelper->isLogged()){
+                $userName= $this->authHelper->getLoggedUserName();
+                $rolUser = $this->authHelper->getLoggedUserRol();
+                $this->view->showBiblioteca($biblioteca,$rolUser, $userName);
+            }else{
+                $this->view->showBiblioteca($biblioteca);
+            }
+            
         }else{
             $mensajeError = "No hay libros en la biblioteca";
-            $this->viewError->showErrorPage($mensajeError);
+            if($this->authHelper->isLogged()){
+                $this->functionHelper->showErrorPage($mensajeError,$this->authHelper->getLoggedUserRol(),$this->authHelper->getLoggedUserName());
+            }else{
+                $this->functionHelper->showErrorPage($mensajeError);
+            }
+            
         }
     }
 
@@ -46,12 +58,27 @@ class LibroController
             if($book){
                 $autor = $this->modelAutor->getAuthor($book->id_autor);
                 if($nombre == $autor->nombre && $apellido == $autor->apellido){
-                    $this->view->showBook($book,$nombre,$apellido);
+                    if($this->authHelper->isLogged()){
+                        $userName= $this->authHelper->getLoggedUserName();
+                        $rolUser = $this->authHelper->getLoggedUserRol();
+                        $userId = $this->authHelper->getLoggedUserId();
+                        $this->view->showBook($book,$nombre,$apellido, $rolUser,$userName,$userId);
+                    }else{
+                        $this->view->showBook($book,$nombre,$apellido);
+                    }
                 }else{
-                    $this->viewError->showErrorPage($mensajeError);
+                    if($this->authHelper->isLogged()){
+                        $this->functionHelper->showErrorPage($mensajeError,$this->authHelper->getLoggedUserRol(),$this->authHelper->getLoggedUserName());
+                    }else{
+                        $this->functionHelper->showErrorPage($mensajeError);
+                    }
                 }
             }else{
-                $this->viewError->showErrorPage($mensajeError);
+                if($this->authHelper->isLogged()){
+                    $this->functionHelper->showErrorPage($mensajeError,$this->authHelper->getLoggedUserRol(),$this->authHelper->getLoggedUserName());
+                }else{
+                    $this->functionHelper->showErrorPage($mensajeError);
+                }
             }
         }
         
@@ -65,14 +92,30 @@ class LibroController
             if($booksOfAuthor){
                 $autor = $this->modelAutor->getAuthor($id);
                 if($nombre == $autor->nombre && $apellido == $autor->apellido){
-                    $this->view->showBooksOfAuthor($booksOfAuthor,$nombre,$apellido);
+
+                    if($this->authHelper->isLogged()){
+                        $userName= $this->authHelper->getLoggedUserName();
+                        $rolUser = $this->authHelper->getLoggedUserRol();
+                        $this->view->showBooksOfAuthor($booksOfAuthor,$nombre,$apellido,$rolUser,$userName);
+                    }else{
+                        $this->view->showBooksOfAuthor($booksOfAuthor,$nombre,$apellido);
+                    }
+
                 }else{
                     $mensajeError = "El autor solicitado no se encuentra en nuestra base de datos";
-                    $this->viewError->showErrorPage($mensajeError);
+                    if($this->authHelper->isLogged()){
+                        $this->functionHelper->showErrorPage($mensajeError,$this->authHelper->getLoggedUserRol(),$this->authHelper->getLoggedUserName());
+                    }else{
+                        $this->functionHelper->showErrorPage($mensajeError);
+                    }
                 }
             }else{
                 $mensajeError = "Ups! Parece que algo salió mal";
-                    $this->viewError->showErrorPage($mensajeError);
+                if($this->authHelper->isLogged()){
+                    $this->functionHelper->showErrorPage($mensajeError,$this->authHelper->getLoggedUserRol(),$this->authHelper->getLoggedUserName());
+                }else{
+                    $this->functionHelper->showErrorPage($mensajeError);
+                }
             }
 
         }   
@@ -80,42 +123,69 @@ class LibroController
     }
 
     function showHomeAdmin(){
-        $this->authHelper->checkLoggedIn();
         $biblioteca = $this->model->getBiblioteca();
         $autores = $this->modelAutor->getAuthors();
-        $this->viewAdmin->showBibliotecaAdmin($biblioteca,$autores);
-    }
-
-    function editarBook($id){
-        $this->authHelper->checkLoggedIn();
-        if(!empty($_POST['titulo'])&& !empty($_POST['autor'])&& !empty($_POST['anio'])){
-            $titulo = $_POST['titulo'];
-            $autor = $_POST['autor'];
-            $anio = $_POST['anio'];
-            $sinopsis = $_POST['sinopsis'];
-            $this->model->updateBook($id,$titulo, $autor,$anio,$sinopsis);
+        
+        if($this->authHelper->isLogged()){
+            $userName= $this->authHelper->getLoggedUserName();
+            $rolUser = $this->authHelper->getLoggedUserRol();
+            if($rolUser=='administrador'){
+                $this->viewAdmin->showBibliotecaAdmin($biblioteca,$autores,$userName,$rolUser);
+            }else{
+                $mensajeError = "Ups! Parece que algo salió mal";
+                $this->functionHelper->showErrorPage($mensajeError,$rolUser,$userName);            }
+        }else{
+            header("Location:".BASE_URL."login");
         }
     }
+
+    // function editarBook($id){
+    //     $this->authHelper->checkLoggedIn();
+    //     if($this->authHelper->getLoggedUserRol()=='administrador' && !empty($_POST['titulo'])&& !empty($_POST['autor'])&& !empty($_POST['anio'])&& !empty($_POST['sinopsis'])){
+    //         $titulo = $_POST['titulo'];
+    //         $autor = $_POST['autor'];
+    //         $anio = $_POST['anio'];
+    //         $sinopsis = $_POST['sinopsis'];
+    //         $this->model->updateBook($id,$titulo, $autor,$anio,$sinopsis);
+    //     }
+    // }
 
     
     function deleteBook($id){
         $this->authHelper->checkLoggedIn();
-        $this->model->deleteBookFromDB($id);
-        header("Location:".BASE_URL."homeAdmin");
-
+        if($this->authHelper->getLoggedUserRol()=='administrador' ){
+            $this->model->deleteBookFromDB($id);
+            header("Location:".BASE_URL."homeAdmin");
+        }else{
+            $mensajeError = "Ups! Parece que algo salió mal";
+            $this->functionHelper->showErrorPage($mensajeError,$this->authHelper->getLoggedUserRol(),$this->authHelper->getLoggedUserName());
+        }
+       
     }
 
     function updateBookForm($id){
-        $this->authHelper->checkLoggedIn();
-        $book = $this->model->getBook($id);
-        $autores = $this->modelAutor->getAuthors();
-        $this->viewAdmin->showUpdateBook($book,$autores);
+        if($this->authHelper->isLogged()){
+            $userName= $this->authHelper->getLoggedUserName();
+            $rolUser = $this->authHelper->getLoggedUserRol();
+            if($rolUser=='administrador'){
+                $book = $this->model->getBook($id);
+                $autores = $this->modelAutor->getAuthors();
+                $this->viewAdmin->showUpdateBook($book,$autores,$userName,$rolUser);
+            }else{
+                $mensajeError = "Ups! Parece que algo salió mal";
+                $this->functionHelper->showErrorPage($mensajeError,$rolUser,$userName);
+            }
+            
+        }else{
+            header("Location:".BASE_URL."login");
+        }
+        
 
     }
 
     function updateBook($id){
         $this->authHelper->checkLoggedIn();
-        if(!empty($_POST['titulo'])&& !empty($_POST['autor'])&& !empty($_POST['anio'])){
+        if($this->authHelper->getLoggedUserRol()=='administrador' && !empty($_POST['titulo'])&& !empty($_POST['autor'])&& !empty($_POST['anio'])){
             $titulo = $_POST['titulo'];
             $autor = $_POST['autor'];
             $anio = $_POST['anio'];
@@ -123,12 +193,15 @@ class LibroController
 
             $this->model->updateBook($id,$titulo, $autor,$anio,$sinopsis);
             header("Location:".BASE_URL."homeAdmin");
+        }else{
+            $mensajeError = "Ups! Parece que algo salió mal";
+            $this->functionHelper->showErrorPage($mensajeError,$this->authHelper->getLoggedUserRol(),$this->authHelper->getLoggedUserName());
         }
     }
 
     function addBook(){
         $this->authHelper->checkLoggedIn();
-        if(!empty($_POST['titulo']) && !empty($_POST['anio'])){
+        if($this->authHelper->getLoggedUserRol()=='administrador' && !empty($_POST['titulo']) && !empty($_POST['anio'])){
             $titulo = $_POST['titulo'];
             $autor = $_POST['autor'];
             $anio = $_POST['anio'];
@@ -136,6 +209,9 @@ class LibroController
 
             $this->model->addBookDB($titulo,$autor,$anio, $sinopsis);
             header("Location:".BASE_URL."homeAdmin");
+        }else{
+            $mensajeError = "Ups! Parece que algo salió mal";
+            $this->functionHelper->showErrorPage($mensajeError,$this->authHelper->getLoggedUserRol(),$this->authHelper->getLoggedUserName());
         }
     }
